Update README.md
An active thread on Reddit is \Random Acts of Pizza” where a user can put in a request for a pizza to be donated by another user. Some requests are granted, others are not. The le pizza_requests.txt includes requests from approximately 5600 pizza requests. Data for each request includes a statement describing the nature of the request, and various other pieces of information. The le README_pizza.txt gives more information about the data set.